refactor(sdk): put evaluator and schema modules at their derived taxonomy path
Both directories were flat except for math, which sat under `math/` -- a third naming
scheme, matching neither its domain (`academic-standards-alignment`) nor its skill
(`mathematics`). Adding the seven feedback evaluators would have made each directory
sixteen modules deep, interleaved with base.ts, inputs.ts and credentials.ts.

A module's path is now derived from its evaluator id, by the same rule the conformance
suite uses to find a contract: dots become separators, underscores become hyphens. So
the location is a fact about the registry rather than a convention, a contract that
moves takes its module with it, and two evaluators sharing a last id segment can no
longer collide -- which schema generation would otherwise have had to work around.

Grouping the evaluators also separates them from the machinery for free: src/evaluators
now holds four infrastructure modules and three domain directories.

The conformance suite asserts the derived path exists for every evaluator. The compiler
cannot: a module moved with its imports fixed typechecks cleanly, and that is exactly
the drift worth catching. Verified by moving one schema up a level, fixing its three
importers, and confirming only this assertion fails.

Moves and import paths only -- no content changes, verified by diffing out every line
that is not an import.

Prompt modules are deliberately left flat: they are near-identical wrappers around
createPromptRenderers and most of them disappear with the single-step factory, so moving
them now would be churn against work that deletes them.
