+++ sdks/typescript/src/evaluators/student-facing-text/ela-reading/organizational-structure.ts
@@ -1,31 +1,34 @@
-import type { LLMProvider } from '../providers/index.js';
-import { PurposeClarityOutputSchema, type PurposeClarityInternal } from '../schemas/purpose-clarity.js';
-import { runPreprocessingStep } from '../features/preprocessing.js';
-import { getSystemPrompt, getUserPrompt } from '../prompts/purpose-clarity/index.js';
-import type { EvaluationResult } from '../schemas/index.js';
-import { BaseEvaluator, Provider, type BaseEvaluatorConfig } from './base.js';
-import { validateInputs, type InputsOf } from './inputs.js';
-import { declaredCredentials } from './credentials.js';
-import type { StageDetail } from '../telemetry/index.js';
-import { EvaluatorError, wrapProviderError } from '../errors.js';
-import CONFIG from '../../../../evals/student-facing-text/ela-reading/purpose-clarity/config.json';
-import INPUT_SCHEMA from '../../../../evals/student-facing-text/ela-reading/purpose-clarity/input_schema.json';
+import type { LLMProvider } from '../../../providers/index.js';
+import {
+  OrganizationalStructureOutputSchema,
+  type OrganizationalStructureInternal,
+} from '../../../schemas/student-facing-text/ela-reading/organizational-structure.js';
+import { runPreprocessingStep } from '../../../features/preprocessing.js';
+import { getSystemPrompt, getUserPrompt } from '../../../prompts/organizational-structure/index.js';
+import type { EvaluationResult } from '../../../schemas/index.js';
+import { BaseEvaluator, Provider, type BaseEvaluatorConfig } from '../../base.js';
+import { validateInputs, type InputsOf } from '../../inputs.js';
+import { declaredCredentials } from '../../credentials.js';
+import type { StageDetail } from '../../../telemetry/index.js';
+import { EvaluatorError, wrapProviderError } from '../../../errors.js';
+import CONFIG from '../../../../../../evals/student-facing-text/ela-reading/organizational-structure/config.json';
+import INPUT_SCHEMA from '../../../../../../evals/student-facing-text/ela-reading/organizational-structure/input_schema.json';
 
 // Step ID convention: "evaluate_{slug}" where slug is the last segment of evaluator.id.
 const STEP_ID = `evaluate_${CONFIG.evaluator.id.split('.').pop()}`;
 const _step = CONFIG.steps.find(s => s.id === STEP_ID);
-if (!_step) throw new Error(`Step "${STEP_ID}" not found in purpose-clarity config.json`);
+if (!_step) throw new Error(`Step "${STEP_ID}" not found in organizational-structure config.json`);
 const STEP = _step;
 
 // Supported grades from input_schema — needed for static metadata, so defined at
 // module level. The enum is the declared set, so it is used verbatim rather than
-// reconstructed from bounds; that admits gaps and non-numeric tokens such as "K".
+// reconstructed from bounds; that also admits non-numeric tokens such as "K".
 const SUPPORTED_GRADES: readonly string[] = INPUT_SCHEMA.properties.grade_level.enum;
 
 /** What this evaluator accepts, taken from its `input_schema.json`. */
-export type PurposeClarityInput = InputsOf<typeof INPUT_SCHEMA>;
+export type OrganizationalStructureInput = InputsOf<typeof INPUT_SCHEMA>;
 
-export class PurposeClarityEvaluator extends BaseEvaluator {
+export class OrganizationalStructureEvaluator extends BaseEvaluator {
   static readonly metadata = {
     id: CONFIG.evaluator.id,
     stableId: CONFIG.evaluator.stable_id,
@@ -42,7 +45,7 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
 
   private static computeFkScore(text: string): number {
     const fkStep = CONFIG.preprocessing.find(p => p.id === 'fk_score');
-    if (!fkStep) throw new Error('fk_score preprocessing step not found in purpose-clarity config.json');
+    if (!fkStep) throw new Error('fk_score preprocessing step not found in organizational-structure config.json');
     return runPreprocessingStep(text, fkStep.implementation.typescript);
   }
 
@@ -57,16 +60,17 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
   }
 
   /**
-   * Evaluate purpose complexity for a given text and grade level
+   * Evaluate organizational structure complexity for a given text and grade level
    *
-   * @param input - The inputs declared in this evaluator's `input_schema.json`
+   * @param text - The text to evaluate
+   * @param gradeLevel - The target grade level (3-12)
    * @returns Evaluation result with complexity score and detailed analysis
-   * @throws {InputValidationError} If an input is missing, unknown, or outside the bounds its schema declares
+   * @throws {InputValidationError} If text is empty, too short/long, or gradeLevel is invalid
    * @throws {ConfigurationError} If modelOverride specifies a model ID that the provider rejects
    * @throws {DependencyError} If the provider call fails (AuthenticationError, RateLimitError, NetworkError, RequestTimeoutError, LLMProviderError)
    * @throws {LLMOutputProcessingError} If the model's response fails its output schema
    */
-  async evaluate(input: PurposeClarityInput): Promise<EvaluationResult<PurposeClarityInternal>> {
+  async evaluate(input: OrganizationalStructureInput): Promise<EvaluationResult<OrganizationalStructureInternal>> {
     let text = '';
     let gradeLevel = '';
     const startTime = Date.now();
@@ -78,14 +82,14 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
       validateInputs(input, INPUT_SCHEMA);
       ({ text, grade_level: gradeLevel } = input);
 
-      this.logger.info('Starting Purpose Clarity evaluation', {
-        evaluator: PurposeClarityEvaluator.metadata.id,
+      this.logger.info('Starting Organizational Structure evaluation', {
+        evaluator: OrganizationalStructureEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
-        textLength: text?.length,
+        textLength: text.length,
       });
 
-      const fkScore = PurposeClarityEvaluator.computeFkScore(text);
+      const fkScore = OrganizationalStructureEvaluator.computeFkScore(text);
       const promptInputs: Record<string, string> = {
         ...input,
         fk_score: String(fkScore),
@@ -105,8 +109,8 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
         token_usage: tokenUsage,
       });
 
-      const result: EvaluationResult<PurposeClarityInternal> = {
-        evaluator: PurposeClarityEvaluator.metadata.id,
+      const result: EvaluationResult<OrganizationalStructureInternal> = {
+        evaluator: OrganizationalStructureEvaluator.metadata.id,
         result: response.data,
         metadata: {
           model: this.provider.label,
@@ -129,8 +133,8 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
         inputText: text,
       }).catch(() => undefined);
 
-      this.logger.info('Purpose Clarity evaluation completed successfully', {
-        evaluator: PurposeClarityEvaluator.metadata.id,
+      this.logger.info('Organizational Structure evaluation completed successfully', {
+        evaluator: OrganizationalStructureEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
         score: response.data.complexity_score,
@@ -141,8 +145,8 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
     } catch (error) {
       const latencyMs = Date.now() - startTime;
 
-      this.logger.error('Purpose Clarity evaluation failed', {
-        evaluator: PurposeClarityEvaluator.metadata.id,
+      this.logger.error('Organizational Structure evaluation failed', {
+        evaluator: OrganizationalStructureEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
         error: error instanceof Error ? error : undefined,
@@ -160,7 +164,7 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
         status: 'error',
         latencyMs,
         textLength: text.length,
-        gradeLevel,
+        gradeLevel: String(gradeLevel),
         provider: this.provider.label,
         tokenUsage,
         errorCode: error instanceof Error ? error.name : 'UnknownError',
@@ -175,23 +179,23 @@ export class PurposeClarityEvaluator extends BaseEvaluator {
 
   private async callLLM(
     inputs: Record<string, string>,
-  ): Promise<{ data: PurposeClarityInternal; usage: { inputTokens: number; outputTokens: number }; latencyMs: number }> {
+  ): Promise<{ data: OrganizationalStructureInternal; usage: { inputTokens: number; outputTokens: number }; latencyMs: number }> {
     const response = await this.provider.generateStructured({
       messages: [
         { role: 'system', content: getSystemPrompt(inputs) },
         { role: 'user', content: getUserPrompt(inputs) },
       ],
-      schema: PurposeClarityOutputSchema,
-      temperature: PurposeClarityEvaluator.TEMPERATURE,
+      schema: OrganizationalStructureOutputSchema,
+      temperature: OrganizationalStructureEvaluator.TEMPERATURE,
     });
 
     return { data: response.data, usage: response.usage, latencyMs: response.latencyMs };
   }
 }
 
-export async function evaluatePurposeClarity(
-  input: PurposeClarityInput,
+export async function evaluateOrganizationalStructure(
+  input: OrganizationalStructureInput,
   config: BaseEvaluatorConfig,
-): Promise<EvaluationResult<PurposeClarityInternal>> {
-  return new PurposeClarityEvaluator(config).evaluate(input);
+): Promise<EvaluationResult<OrganizationalStructureInternal>> {
+  return new OrganizationalStructureEvaluator(config).evaluate(input);
 }
