+++ sdks/typescript/src/evaluators/student-facing-text/ela-reading/purpose-clarity.ts
@@ -1,34 +1,31 @@
-import type { LLMProvider } from '../providers/index.js';
-import {
-  OrganizationalStructureOutputSchema,
-  type OrganizationalStructureInternal,
-} from '../schemas/organizational-structure.js';
-import { runPreprocessingStep } from '../features/preprocessing.js';
-import { getSystemPrompt, getUserPrompt } from '../prompts/organizational-structure/index.js';
-import type { EvaluationResult } from '../schemas/index.js';
-import { BaseEvaluator, Provider, type BaseEvaluatorConfig } from './base.js';
-import { validateInputs, type InputsOf } from './inputs.js';
-import { declaredCredentials } from './credentials.js';
-import type { StageDetail } from '../telemetry/index.js';
-import { EvaluatorError, wrapProviderError } from '../errors.js';
-import CONFIG from '../../../../evals/student-facing-text/ela-reading/organizational-structure/config.json';
-import INPUT_SCHEMA from '../../../../evals/student-facing-text/ela-reading/organizational-structure/input_schema.json';
+import type { LLMProvider } from '../../../providers/index.js';
+import { PurposeClarityOutputSchema, type PurposeClarityInternal } from '../../../schemas/student-facing-text/ela-reading/purpose-clarity.js';
+import { runPreprocessingStep } from '../../../features/preprocessing.js';
+import { getSystemPrompt, getUserPrompt } from '../../../prompts/purpose-clarity/index.js';
+import type { EvaluationResult } from '../../../schemas/index.js';
+import { BaseEvaluator, Provider, type BaseEvaluatorConfig } from '../../base.js';
+import { validateInputs, type InputsOf } from '../../inputs.js';
+import { declaredCredentials } from '../../credentials.js';
+import type { StageDetail } from '../../../telemetry/index.js';
+import { EvaluatorError, wrapProviderError } from '../../../errors.js';
+import CONFIG from '../../../../../../evals/student-facing-text/ela-reading/purpose-clarity/config.json';
+import INPUT_SCHEMA from '../../../../../../evals/student-facing-text/ela-reading/purpose-clarity/input_schema.json';
 
 // Step ID convention: "evaluate_{slug}" where slug is the last segment of evaluator.id.
 const STEP_ID = `evaluate_${CONFIG.evaluator.id.split('.').pop()}`;
 const _step = CONFIG.steps.find(s => s.id === STEP_ID);
-if (!_step) throw new Error(`Step "${STEP_ID}" not found in organizational-structure config.json`);
+if (!_step) throw new Error(`Step "${STEP_ID}" not found in purpose-clarity config.json`);
 const STEP = _step;
 
 // Supported grades from input_schema — needed for static metadata, so defined at
 // module level. The enum is the declared set, so it is used verbatim rather than
-// reconstructed from bounds; that also admits non-numeric tokens such as "K".
+// reconstructed from bounds; that admits gaps and non-numeric tokens such as "K".
 const SUPPORTED_GRADES: readonly string[] = INPUT_SCHEMA.properties.grade_level.enum;
 
 /** What this evaluator accepts, taken from its `input_schema.json`. */
-export type OrganizationalStructureInput = InputsOf<typeof INPUT_SCHEMA>;
+export type PurposeClarityInput = InputsOf<typeof INPUT_SCHEMA>;
 
-export class OrganizationalStructureEvaluator extends BaseEvaluator {
+export class PurposeClarityEvaluator extends BaseEvaluator {
   static readonly metadata = {
     id: CONFIG.evaluator.id,
     stableId: CONFIG.evaluator.stable_id,
@@ -45,7 +42,7 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
 
   private static computeFkScore(text: string): number {
     const fkStep = CONFIG.preprocessing.find(p => p.id === 'fk_score');
-    if (!fkStep) throw new Error('fk_score preprocessing step not found in organizational-structure config.json');
+    if (!fkStep) throw new Error('fk_score preprocessing step not found in purpose-clarity config.json');
     return runPreprocessingStep(text, fkStep.implementation.typescript);
   }
 
@@ -60,17 +57,16 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
   }
 
   /**
-   * Evaluate organizational structure complexity for a given text and grade level
+   * Evaluate purpose complexity for a given text and grade level
    *
-   * @param text - The text to evaluate
-   * @param gradeLevel - The target grade level (3-12)
+   * @param input - The inputs declared in this evaluator's `input_schema.json`
    * @returns Evaluation result with complexity score and detailed analysis
-   * @throws {InputValidationError} If text is empty, too short/long, or gradeLevel is invalid
+   * @throws {InputValidationError} If an input is missing, unknown, or outside the bounds its schema declares
    * @throws {ConfigurationError} If modelOverride specifies a model ID that the provider rejects
    * @throws {DependencyError} If the provider call fails (AuthenticationError, RateLimitError, NetworkError, RequestTimeoutError, LLMProviderError)
    * @throws {LLMOutputProcessingError} If the model's response fails its output schema
    */
-  async evaluate(input: OrganizationalStructureInput): Promise<EvaluationResult<OrganizationalStructureInternal>> {
+  async evaluate(input: PurposeClarityInput): Promise<EvaluationResult<PurposeClarityInternal>> {
     let text = '';
     let gradeLevel = '';
     const startTime = Date.now();
@@ -82,14 +78,14 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
       validateInputs(input, INPUT_SCHEMA);
       ({ text, grade_level: gradeLevel } = input);
 
-      this.logger.info('Starting Organizational Structure evaluation', {
-        evaluator: OrganizationalStructureEvaluator.metadata.id,
+      this.logger.info('Starting Purpose Clarity evaluation', {
+        evaluator: PurposeClarityEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
-        textLength: text.length,
+        textLength: text?.length,
       });
 
-      const fkScore = OrganizationalStructureEvaluator.computeFkScore(text);
+      const fkScore = PurposeClarityEvaluator.computeFkScore(text);
       const promptInputs: Record<string, string> = {
         ...input,
         fk_score: String(fkScore),
@@ -109,8 +105,8 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
         token_usage: tokenUsage,
       });
 
-      const result: EvaluationResult<OrganizationalStructureInternal> = {
-        evaluator: OrganizationalStructureEvaluator.metadata.id,
+      const result: EvaluationResult<PurposeClarityInternal> = {
+        evaluator: PurposeClarityEvaluator.metadata.id,
         result: response.data,
         metadata: {
           model: this.provider.label,
@@ -133,8 +129,8 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
         inputText: text,
       }).catch(() => undefined);
 
-      this.logger.info('Organizational Structure evaluation completed successfully', {
-        evaluator: OrganizationalStructureEvaluator.metadata.id,
+      this.logger.info('Purpose Clarity evaluation completed successfully', {
+        evaluator: PurposeClarityEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
         score: response.data.complexity_score,
@@ -145,8 +141,8 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
     } catch (error) {
       const latencyMs = Date.now() - startTime;
 
-      this.logger.error('Organizational Structure evaluation failed', {
-        evaluator: OrganizationalStructureEvaluator.metadata.id,
+      this.logger.error('Purpose Clarity evaluation failed', {
+        evaluator: PurposeClarityEvaluator.metadata.id,
         operation: 'evaluate',
         gradeLevel,
         error: error instanceof Error ? error : undefined,
@@ -164,7 +160,7 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
         status: 'error',
         latencyMs,
         textLength: text.length,
-        gradeLevel: String(gradeLevel),
+        gradeLevel,
         provider: this.provider.label,
         tokenUsage,
         errorCode: error instanceof Error ? error.name : 'UnknownError',
@@ -179,23 +175,23 @@ export class OrganizationalStructureEvaluator extends BaseEvaluator {
 
   private async callLLM(
     inputs: Record<string, string>,
-  ): Promise<{ data: OrganizationalStructureInternal; usage: { inputTokens: number; outputTokens: number }; latencyMs: number }> {
+  ): Promise<{ data: PurposeClarityInternal; usage: { inputTokens: number; outputTokens: number }; latencyMs: number }> {
     const response = await this.provider.generateStructured({
       messages: [
         { role: 'system', content: getSystemPrompt(inputs) },
         { role: 'user', content: getUserPrompt(inputs) },
       ],
-      schema: OrganizationalStructureOutputSchema,
-      temperature: OrganizationalStructureEvaluator.TEMPERATURE,
+      schema: PurposeClarityOutputSchema,
+      temperature: PurposeClarityEvaluator.TEMPERATURE,
     });
 
     return { data: response.data, usage: response.usage, latencyMs: response.latencyMs };
   }
 }
 
-export async function evaluateOrganizationalStructure(
-  input: OrganizationalStructureInput,
+export async function evaluatePurposeClarity(
+  input: PurposeClarityInput,
   config: BaseEvaluatorConfig,
-): Promise<EvaluationResult<OrganizationalStructureInternal>> {
-  return new OrganizationalStructureEvaluator(config).evaluate(input);
+): Promise<EvaluationResult<PurposeClarityInternal>> {
+  return new PurposeClarityEvaluator(config).evaluate(input);
 }
