+++ sdks/typescript/src/schemas/student-facing-text/ela-reading/purpose-clarity.ts
